# docs/swagger.yaml
basePath: /
definitions:
...
  sites.Site:
    properties:
      host:
        example: site.com
        type: string
      name:
        example: site
        type: string
    type: object
host: localhost:80
info:
  contact: {}
  description: How to work CRUD server in reverseProxy
  title: CRUD server in reverseProxy
  version: 1.0.0
paths:
...
  /sites:
    post:
      consumes:
      - application/json
      description: Create site
      parameters:
      - description: site info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/sites.Site'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: string
      summary: Create new site
      tags:
      - Sites
  /sites/{id}:
    delete:
      consumes:
      - application/json
      description: delete site
      parameters:
      - description: site ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sites.Site'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete site based on given id
      tags:
      - Sites
    get:
      consumes:
      - application/json
      description: get site
      parameters:
      - description: site ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sites.Site'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get site based on given id
      tags:
      - Sites
    put:
      consumes:
      - application/json
      description: update site
      parameters:
      - description: site ID
        in: path
        name: id
        required: true
        type: integer
      - description: site info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/sites.Site'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sites.Site'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Update site based on given id
      tags:
      - Sites
schemes:
- http
swagger: "2.0"